feat: move Voodoo types to voodoo/types.ts
