+++ src/voodoo/conversations/VoodooConversations.ts
@@ -1,10 +1,11 @@
 import { messageApi } from '@xmtp/proto'
 import { Mutex } from 'async-mutex'
-import VoodooClient, {
+import VoodooClient from '../VoodooClient'
+import {
   VoodooContact,
   EncryptedVoodooMessage,
   VoodooMessage,
-} from '../VoodooClient'
+} from '../types'
 import VoodooConversation from './VoodooConversation'
 import { utils } from '../../crypto'
